Move pkg/instrumentation to internal/instrumentation
